Add A-packages
