+++ src/main/java/Duke/Exceptions/ipException.java
@@ -1,3 +1,7 @@
+package Duke.Exceptions;
+
+import Duke.Duke;
+
 public class ipException extends Exception {
 
 }
