+++ src/main/java/Duke/Exceptions/ipException2.java
@@ -1,3 +1,7 @@
+package Duke.Exceptions;
+
+import Duke.Duke;
+
 public class ipException2 extends Exception {
 
 }
