+++ src/main/java/Duke/Exceptions/ipException3.java
@@ -1,3 +1,7 @@
+package Duke.Exceptions;
+
+import Duke.Duke;
+
 public class ipException3 extends Exception {
 
 }
